login, sginup and reset password api added.

+++ src/main/java/com/journal/entity/FileRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by Venturedive on 10/29/2017.
  */
-public interface FileRepository extends CrudRepository<File, Long> {
+public interface FileRepository extends CrudRepository<File, Integer> {
 
 }
