+++ src/main/java/com/journal/entity/UserRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by Venturedive on 10/29/2017.
  */
-public interface UserRepository extends CrudRepository<User, Long> {
-
+public interface UserRepository extends CrudRepository<User, Integer> {
 }
